reorganize app fixtures
